+++ test/prufer_test.cpp
@@ -27,10 +27,10 @@ vector<pair<int, int>> random_tree(int size, int offset) {
         roots.push_back(i + offset);
 
     while (roots.size() > 1) {
-        int x = uni(0, roots.size() - 1)(gen);
+        int x = uni(0, (int)roots.size() - 1)(gen);
         swap(roots[0], roots[x]);
         x = 0;
-        int y = uni(1, roots.size() - 1)(gen);
+        int y = uni(1, (int)roots.size() - 1)(gen);
         swap(roots[1], roots[y]);
         y = 1;
